pyproject.toml
[tool.poetry]
name = "define-json"
version = "0.1.0"
description = "Define-JSON CDISC standard for data contracts"
authors = ["Jeremy Teoh"]
readme = "README.md"
package-mode = true
packages = [{include = "define_json", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
linkml = "^1.9.2"
linkml-runtime = "^1.9.3"
pyyaml = "^6.0.2"
jsonschema = "^4.24.0"
click = "^8.2.1"
rich = "^14.0.0"
plotly = "^6.2.0"
lxml = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
pandas = "^2.3.0"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
yamllint = "^1.37.1"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.15"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
mkdocs-mermaid2-plugin = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

